add formatDate func and export it w formatTime in utils; create consistency for date and time formats
